tests/requirements-test.txt
# Testing requirements for CodeWeaver

# Core testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-xdist>=3.3.0  # For parallel test execution

# Web testing
pytest-aiohttp>=1.0.4
aioresponses>=0.7.4

# Additional testing utilities
factory-boy>=3.3.0  # For creating test data
freezegun>=1.2.2    # For mocking datetime
responses>=0.23.0   # For mocking HTTP requests

# Coverage and reporting
coverage>=7.2.0
pytest-html>=3.2.0  # HTML test reports
pytest-json-report>=1.5.0  # JSON test reports

# Performance testing
pytest-benchmark>=4.0.0

# For testing CLI applications
click-testing>=1.0.0

# Memory profiling in tests
memory-profiler>=0.61.0

# Fixtures and test data
pytest-fixtures>=0.1.0

# Mock external services
responses>=0.23.0
httpretty>=1.1.4

# Database testing (if needed)
sqlalchemy-utils>=0.41.0

# Temporary file and directory utilities
tempfile-extra>=0.1.0

# For testing with real files and directories
pyfakefs>=5.2.0  # Fake filesystem for testing
